[BUGFIX] Prevent Aloha from destroying RTE links
Since Aloha is Frontend editing, it works directly with the transformed
source code. Links set in the RTE are stored in a proprietary <link>
tag, e.g.

<link 123 500x500>Link to internal page 123, opens in popup</link>
<link file:123>Link to file with uid 123</link>
<link http://typo3.org>Link to external website</link>

These tags are transformed to real links that don't make it possible
to be "reverse-engineered":

<a href="my/page/123" onclick="popup code">[...]
<a href="/fileadmin/document.pdf">[...]
<a href="http://typo3.org">[...]

Therefore, as soon as a content element containing typolinks is saved,
the links are destroyed. In previous versions, this was tackled by
deactivating RealURL which kept internal and external links working.
But this does not work for linkhandler links (such as file:123) and
for additional parameters.

How to make it work:

The RTE parseFunc (lib.parseFunc_RTE) which defines the parsing for all
typolinks from RTE is extended to add the original typolink parameters
in a custom attribute (data-typolinkoriginalparameter) to each link.
The link is then output like this:

<a href="fileadmin/file.pdf" data-typolinkoriginalparameter="file:123">

Since the Aloha contentHandler removes all attributes that are not
defined explicitely, the Aloha configuration is changed to allow this
attribute.

On saving, the content is passed through the CeRteLinks
RequestPreProcess hook. In this hook, each link is parsed and replaced
by an original typolink (<link> tag). Then the content is parsed with
RteHtmlParser::TS_links_rte that converts <link> to <a> tags.

This format is now accepted by the DataHandler (former TCEmain)
for updating the record and keeping all the links.