[TASK] Deactivate Aloha in Workspaces Preview
Aloha is not respecting workspaces, but is currently loaded in the
Workspaces Preview. Apart from the fact that the Aloha toolbar is
covered by the Workspace Preview toolbar, all changes made in the
Workspaces Preview would be directly saved to the live version.
Therefore Aloha must not be activated in the Workspaces Preview.