[BUGFIX] Don't disable save button after changing a CE and entering second CE
When a CE is edited in Aloha and then left, the information about the
elements to save and the save button are displayed. On entering another
CE, the save button is hidden so it is no longer possible to save the
content element without doing another change.

This change checks for unsaved content and doesn't hide the save button
if there is unsaved content. The same change also hides the save and
discard button on launching Aloha because there can't be anything to
save at this time.